[Refactor] CreatePost에서 사용하는 Input 컴포넌트 리펙터링 #261
